+++ proto/c/src/soil_power_sensor.pb.c
@@ -15,6 +15,9 @@ PB_BIND(PowerMeasurement, PowerMeasurement, AUTO)
 PB_BIND(Teros12Measurement, Teros12Measurement, AUTO)
 
 
+PB_BIND(Teros21Measurement, Teros21Measurement, AUTO)
+
+
 PB_BIND(Phytos31Measurement, Phytos31Measurement, AUTO)
 
 
@@ -49,12 +52,15 @@ PB_BIND(UserConfiguration, UserConfiguration, AUTO)
 
 
 
+<<<<<<< HEAD
 
 
 
 
 
 
+=======
+>>>>>>> main
 #ifndef PB_CONVERT_DOUBLE_FLOAT
 /* On some platforms (such as AVR), double is really float.
  * To be able to encode/decode double on these platforms, you need.