+++ stm32/Src/main.c
@@ -41,9 +41,10 @@
 #include "rtc.h"
 #include "sensors.h"
 #include "wifi.h"
-
 #include "controller/controller.h"
 #include "controller/wifi.h"
+#include "userConfig.h"
+#include "teros21.h"
 /* USER CODE END Includes */
 
 /* Private typedef -----------------------------------------------------------*/
@@ -78,6 +79,8 @@
 
 /* USER CODE END 0 */
 
+void SystemClock_Config(void);
+
 /**
   * @brief  The application entry point.
   * @retval int
@@ -109,7 +112,26 @@ int main(void)
   MX_DMA_Init();
   MX_ADC_Init();
   MX_USART1_UART_Init();
-  MX_I2C2_Init();
+  MX_I2C2_Init(); 
+  SystemApp_Init();
+
+  if (UserConfigLoad() != USERCONFIG_OK) {
+    APP_LOG(TS_OFF, VLEVEL_M, "Error loading user configuration!\n");
+    APP_LOG(TS_OFF, VLEVEL_M, "Waiting for new configuration...\n");
+
+    // TODO implement status code with LED
+
+    // Wait for new configuration
+    //UserConfig_ProcessDataPolling();
+    UserConfig_InitAdvanceTrace();
+    while (1);
+  }
+
+  UserConfigPrint();
+
+  MX_LoRaWAN_Init();
+
+  // required for SDI-12
   MX_USART2_UART_Init();
   MX_TIM1_Init();
   /* USER CODE BEGIN 2 */
@@ -119,17 +141,24 @@ int main(void)
   
   // init external adc
   ADC_init();
-
-  // initialize
+  //MX_RTC_Init();
   SensorsInit();
-  FIFO_Init();
+
+  // initialize the user config interrupt
+  UserConfig_InitAdvanceTrace();
+
+  // alternative blocking polling method
+  //UserConfig_ProcessDataPolling();
+
+  // currently not functional
+  //FIFO_Init();
 
   // Debug message, gets printed after init code
   APP_PRINTF("Soil Power Sensor Wio-E5 firmware, compiled on %s %s\n", __DATE__, __TIME__);
 
   // configure sensors
   //SensorsAdd(SensorsMeasureTest);
-  SensorsAdd(ADC_measure);
+  //SensorsAdd(ADC_measure);
   //SensorsAdd(SDI12_Teros12Measure);
   //SensorsAdd(Phytos31_measure);
   
@@ -141,10 +170,37 @@ int main(void)
   ControllerInit();
 
   // init either WiFi or LoRaWAN
+  // TODO add configuration for upload method
   //MX_LoRaWAN_Init();
   WiFiInit();
 
 
+  // get the current user config
+  const UserConfiguration* cfg = UserConfigGet();
+ 
+
+  // configure enabled sensors
+  for (int i=0; i < cfg->enabled_sensors_count; i++) {
+    EnabledSensor sensor = cfg->enabled_sensors[i];
+    if ((sensor == EnabledSensor_Voltage) || (sensor == EnabledSensor_Current)) {
+      SensorsAdd(ADC_measure);
+      APP_LOG(TS_ON, VLEVEL_M, "ADS Enabled!\n");
+    }
+    if (sensor == EnabledSensor_Teros12) {
+      APP_LOG(TS_ON, VLEVEL_M, "Teros12 not implemented!\n");
+      //SensorsAdd(SDI12_Teros12Measure);
+    }
+    if (sensor == EnabledSensor_BME280) {
+      BME280Init();
+      SensorsAdd(BME280Measure);
+      APP_LOG(TS_ON, VLEVEL_M, "BME280 Enabled!\n");
+    }
+    if (sensor == EnabledSensor_Teros21) {
+      SensorsAdd(Teros21Measure);
+      APP_LOG(TS_ON, VLEVEL_M, "Teros21 Enabled!\n");
+    }
+  }
+  
   /* USER CODE END 2 */
 
   /* Infinite loop */
@@ -158,6 +214,11 @@ int main(void)
   }
   /* USER CODE END 3 */
 }
+
+
+/* USER CODE BEGIN 4 */
+
+/* USER CODE END 4 */
 
 /**
   * @brief System Clock Configuration
@@ -179,12 +240,18 @@ void SystemClock_Config(void)
 
   /** Initializes the CPU, AHB and APB buses clocks
   */
-  RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_LSE|RCC_OSCILLATORTYPE_MSI;
+  RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_LSE | RCC_OSCILLATORTYPE_MSI;
   RCC_OscInitStruct.LSEState = RCC_LSE_ON;
   RCC_OscInitStruct.MSIState = RCC_MSI_ON;
   RCC_OscInitStruct.MSICalibrationValue = RCC_MSICALIBRATION_DEFAULT;
-  RCC_OscInitStruct.MSIClockRange = RCC_MSIRANGE_11;
-  RCC_OscInitStruct.PLL.PLLState = RCC_PLL_NONE;
+  RCC_OscInitStruct.MSIClockRange = RCC_MSIRANGE_6;
+  RCC_OscInitStruct.PLL.PLLState = RCC_PLL_ON;
+  RCC_OscInitStruct.PLL.PLLSource = RCC_PLLSOURCE_MSI;
+  RCC_OscInitStruct.PLL.PLLM = RCC_PLLM_DIV1;
+  RCC_OscInitStruct.PLL.PLLN = 42;
+  RCC_OscInitStruct.PLL.PLLP = RCC_PLLP_DIV2;
+  RCC_OscInitStruct.PLL.PLLR = RCC_PLLR_DIV4;
+  RCC_OscInitStruct.PLL.PLLQ = RCC_PLLQ_DIV2;
   if (HAL_RCC_OscConfig(&RCC_OscInitStruct) != HAL_OK)
   {
     Error_Handler();
@@ -192,12 +259,10 @@ void SystemClock_Config(void)
 
   /** Configure the SYSCLKSource, HCLK, PCLK1 and PCLK2 clocks dividers
   */
-  RCC_ClkInitStruct.ClockType = RCC_CLOCKTYPE_HCLK3|RCC_CLOCKTYPE_HCLK
-                              |RCC_CLOCKTYPE_SYSCLK|RCC_CLOCKTYPE_PCLK1
-                              |RCC_CLOCKTYPE_PCLK2;
-  RCC_ClkInitStruct.SYSCLKSource = RCC_SYSCLKSOURCE_MSI;
-  RCC_ClkInitStruct.AHBCLKDivider = RCC_SYSCLK_DIV1;
-  RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV1;
+  RCC_ClkInitStruct.ClockType = RCC_CLOCKTYPE_HCLK3 | RCC_CLOCKTYPE_HCLK | RCC_CLOCKTYPE_SYSCLK | RCC_CLOCKTYPE_PCLK1 | RCC_CLOCKTYPE_PCLK2;
+  RCC_ClkInitStruct.SYSCLKSource = RCC_SYSCLKSOURCE_PLLCLK;
+  RCC_ClkInitStruct.AHBCLKDivider = RCC_SYSCLK_DIV5;
+  RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV4;
   RCC_ClkInitStruct.APB2CLKDivider = RCC_HCLK_DIV1;
   RCC_ClkInitStruct.AHBCLK3Divider = RCC_SYSCLK_DIV1;
 
@@ -207,10 +272,6 @@ void SystemClock_Config(void)
   }
   HAL_RCC_MCOConfig(RCC_MCO1, RCC_MCO1SOURCE_SYSCLK, RCC_MCODIV_1);
 }
-
-/* USER CODE BEGIN 4 */
-
-/* USER CODE END 4 */
 
 /**
   * @brief  This function is executed in case of error occurrence.