+++ stm32/lib/phytos31/src/phytos31.c
@@ -28,7 +28,7 @@ phytos_measurments Phytos31GetMeasurment() {
   return measurments;
 }
 
-size_t Phytos31_measure(uint8_t *data) {
+size_t Phytos31_measure(uint8_t* data) {
   // get timestamp
   SysTime_t ts = SysTimeGet();
   phytos_measurments measurment;
@@ -37,9 +37,11 @@ size_t Phytos31_measure(uint8_t *data) {
   measurment = Phytos31GetMeasurment();
   double adc_voltage_float = measurment.phytos31_raw / 1000.;
 
+  const UserConfiguration* cfg = UserConfigGet();
+
   // encode measurement
-  size_t data_len = EncodePhytos31Measurement(ts.Seconds, LOGGER_ID, CELL_ID,
-                                              adc_voltage_float, 0.0, data);
+  size_t data_len = EncodePhytos31Measurement(
+      ts.Seconds, cfg->logger_id, cfg->cell_id, adc_voltage_float, 0.0, data);
 
   // return number of bytes in serialized measurement
   return data_len;