+++ stm32/test/test_proto/test_proto.c
@@ -295,7 +295,7 @@ int main(void) {
   MX_USART1_UART_Init();
 
   // wait for UART
-  for (int i = 0; i < 1000000; i++) {
+  for (int i = 0; i < 10000000; i++) {
     __NOP();
   }